feat: configuring database - WIP.
