+++ Backend/TaskVisualizerWeb/TaskVisualizerWeb/TaskVisualizerWeb.csproj
@@ -13,4 +13,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\TaskVisualizerWeb.Domain\TaskVisualizerWeb.Domain.csproj" />
+  </ItemGroup>
+
 </Project>
