+++ Backend/TaskVisualizerWeb/TaskVisualizerWeb/appsettings.Development.json
@@ -4,5 +4,8 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
+  },
+  "ConnectionStrings": {
+    "WebApiDatabase": "Host=localhost; Database=dotnet-6-crud-api; Username=ps_user; Password=SecurePassword"
   }
 }
